Adjust cond.prob to match observed levels
Adjust the probability of condom use among men not paired with always condom users such that the overall level of condom use still matches that observed in the data. This adjustment is needed b/c when a man who does not always use condoms is paired with an always user, he is set to use condoms, which affects the overall mean level of condom use.