Revise PrEP coverage, uptake, and eligibility
Because the model does not reproduce observed prevalence by region, do not specify differential PrEP uptake by region. Add scenario for lower PrEP coverage. Also allow men to initialize PrEP as long as they are not diagnosed positive (previously had to be negative). In the first time step of PrEP, start the number of men on PrEP needed to reach initial coverage (i.e. don't only start on PrEP if tested in that time step)